cleanup: moved report field options to their own file
